+++ R/guide_coloursteps_interactive.R
@@ -11,43 +11,44 @@
 #' @example examples/scale_viridis_guide_coloursteps_interactive.R
 #' @seealso [interactive_parameters], [girafe()]
 #' @export
-guide_coloursteps_interactive <- function(...)
-  guide_interactive(guide_coloursteps, "interactive_coloursteps", ...)
+guide_coloursteps_interactive <- function(...) {
+  guide_interactive(guide_coloursteps, ..., interactive_guide = GuideInteractiveColoursteps)
+}
 
 #' @export
 #' @rdname guide_coloursteps_interactive
 guide_colorsteps_interactive <- guide_coloursteps_interactive
 
+#' @rdname ggiraph-ggproto
+#' @format NULL
+#' @usage NULL
 #' @export
-#' @importFrom purrr imap
-guide_train.interactive_coloursteps <- function(guide,
-                                                scale,
-                                                aesthetic = NULL) {
-  zz <- NextMethod()
-  if (is.null(zz))
-    return(zz)
-
-  # just copy them from scale to trained guide
-  ipar <- get_ipar(scale)
-  data <- copy_interactive_attrs(scale, list(), ipar = ipar)
-  zz$.interactive <- data
-  zz$.ipar <- ipar
-  zz
-}
-
-#' @export
-#' @importFrom purrr compact
-guide_gengrob.interactive_coloursteps <- function(guide, theme) {
-  guide_gtable <- NextMethod()
-  ipar = get_ipar(guide)
-  data <- get_interactive_data(guide)
-  # set them to the bar
-  barIndex <- which(guide_gtable$layout$name == "bar")
-  guide_gtable$grobs[[barIndex]] <-
-    add_interactive_attrs(guide_gtable$grobs[[barIndex]],
-                          data,
-                          data_attr = "key-id",
-                          ipar = ipar)
-
-  guide_gtable
-}
+GuideInteractiveColoursteps <- ggproto(
+  "GuideInteractiveColoursteps", GuideColoursteps,
+  train = function(self, params = self$params, scale, aesthetic = NULL, ...) {
+    parent <- ggproto_parent(GuideColoursteps, self)
+    params <- parent$train(params = params, scale = scale, aesthetic = aesthetic, ...)
+    if (!is.null(params) && is.data.frame(params$key) && nrow(params$key)) {
+      parsed <- interactive_guide_parse_binned_breaks(scale, params)
+      breaks <- parsed$all_breaks
+      label_breaks <- parsed$breaks
+      if (params$even.steps || !is.numeric(parsed$scale_breaks)) {
+        show.limits <- params$show.limits %||% scale$show.limits %||% FALSE
+        if (show.limits && !(is.character(scale$labels) || is.numeric(scale$labels))) {
+          label_breaks <- parsed$all_breaks
+        }
+      }
+      params <- interactive_guide_train(params, scale, breaks,
+        label_breaks = label_breaks, max_len = length(breaks) - 1
+      )
+    }
+    params
+  },
+  override_elements = function(params, elements, theme) {
+    elements <- GuideColoursteps$override_elements(params, elements, theme)
+    interactive_guide_override_elements(elements)
+  },
+  build_decor = function(decor, grobs, elements, params) {
+    GuideInteractiveColourbar$build_decor(decor, grobs, elements, params)
+  }
+)