+++ R/ipar.R
@@ -61,25 +61,24 @@
 #' the type of the guide used.
 #' The guides do not accept any interactive parameter directly, they receive them from the scales.
 #'
-#' \itemize{
-#'   \item When guide of type `legend` or `bins` is used, it will be converted to a
-#'   [guide_legend_interactive()] or [guide_bins_interactive()] respectively,
-#'   if it's not already.
+#' When guide of type `legend`, `bins`, `colourbar` or `coloursteps` is used,
+#' it will be converted to a [guide_legend_interactive()], [guide_bins_interactive()],
+#' [guide_colourbar_interactive()] or [guide_coloursteps_interactive()] respectively,
+#' if it's not already.
 #'
-#'   The length of each scale interactive parameter vector should match the length of the breaks.
-#'   It can also be a named vector, where each name should correspond to the same break name.
-#'   It can also be defined as function that takes the breaks as input and returns a named or
-#'   unnamed vector of values as output.
+#' The length of each scale interactive parameter vector should match the length of the breaks.
+#' It can also be a named vector, where each name should correspond to the same break name.
+#' It can also be defined as function that takes the breaks as input and returns a named or
+#' unnamed vector of values as output.
 #'
-#'   The interactive parameters here, give interactivity only to the key elements of the guide.
+#' For binned guides like `bins` and `coloursteps` the breaks include the label breaks and the limits.
+#' The number of bins will be one less than the number of breaks and the interactive parameters can be
+#' constructed for each bin separately (look at the examples).
+#' For `colourbar` guide in raster mode, the breaks vector, is scalar 1 always, meaning the interactive
+#' parameters should be scalar too. For `colourbar` guide in non-raster mode, the bar is drawn using
+#' rectangles, and the breaks are the midpoints of each rectangle.
 #'
-#'   \item When guide of type `colourbar` or `coloursteps` is used, it will be converted to a
-#'   [guide_colourbar_interactive()] or [guide_coloursteps_interactive()]
-#'   respectively, if it's not already.
-#'
-#'   The scale interactive parameters in this case should be scalar values and give interactivity
-#'   to the colorbar only.
-#' }
+#' The interactive parameters here, give interactivity only to the key elements of the guide.
 #'
 #' To provide interactivity to the rest of the elements of a guide, (title, labels, background, etc),
 #' the relevant theme elements or relevant guide arguments can be used.
@@ -229,8 +228,9 @@ add_interactive_attrs <- function(gr,
   anames <- Filter(x = get_interactive_attr_names(data, ipar = ipar), function(a) {
     !is.null(data[[a]])
   })
-  if (length(anames) == 0)
+  if (length(anames) == 0) {
     return(gr)
+  }
 
   # if passed grob is a gTree, loop through the children
   # note that some grobs (like labelgrob) inherit from gTree,
@@ -253,7 +253,6 @@ add_interactive_attrs <- function(gr,
             ipar = anames
           )
       }
-
     } else if (children_len == data_len) {
       # pass the correct data row
       for (i in seq_along(gr$children)) {
@@ -272,7 +271,6 @@ add_interactive_attrs <- function(gr,
       abort("Can't add interactive attrs to gTree", call = NULL)
     }
     return(gr)
-
   } else {
     do_add_interactive_attrs(
       gr = gr,
@@ -295,22 +293,24 @@ do_add_interactive_attrs <- function(gr,
                                      overwrite = TRUE,
                                      data_attr = "data-id",
                                      ipar = IPAR_NAMES) {
-
   # check that is a grob
-  if (!is.grob(gr) || is.zero(gr))
+  if (!is.grob(gr) || is.zero(gr)) {
     return(gr)
+  }
   # check if it's interactive grob already
   isInteractive <- length(grep("interactive_", class(gr))) > 0
   ip <- get_interactive_data(gr)
   if (length(rows) == 0) {
     for (a in ipar) {
-      if (!isInteractive || isTRUE(overwrite) || is.null(ip[[a]]))
+      if (!isInteractive || isTRUE(overwrite) || is.null(ip[[a]])) {
         ip[[a]] <- data[[a]]
+      }
     }
   } else {
     for (a in ipar) {
-      if (!isInteractive || isTRUE(overwrite) || is.null(ip[[a]]))
+      if (!isInteractive || isTRUE(overwrite) || is.null(ip[[a]])) {
         ip[[a]] <- data[[a]][rows]
+      }
     }
   }
   gr$.ipar <- ipar
@@ -330,11 +330,11 @@ do_add_interactive_attrs <- function(gr,
 }
 
 get_interactive_data <- function(x, default = list()) {
-  (if(!is.atomic(x)) x$.interactive) %||% attr(x, "interactive") %||% default
+  (if (!is.atomic(x)) x$.interactive) %||% attr(x, "interactive") %||% default
 }
 
 get_ipar <- function(x, default = IPAR_NAMES) {
-  ipar <- (if(!is.atomic(x)) x$.ipar) %||% attr(x, "ipar")
+  ipar <- (if (!is.atomic(x)) x$.ipar) %||% attr(x, "ipar")
   if (length(ipar) > 0 && is.character(ipar)) {
     ipar
   } else {
@@ -343,7 +343,7 @@ get_ipar <- function(x, default = IPAR_NAMES) {
 }
 
 get_data_attr <- function(x, default = "data-id") {
-  data_attr <- (if(!is.atomic(x)) x$.data_attr) %||% attr(x, "data_attr")
+  data_attr <- (if (!is.atomic(x)) x$.data_attr) %||% attr(x, "data_attr")
   if (length(data_attr) == 1 && is.character(data_attr)) {
     data_attr
   } else {
@@ -356,8 +356,9 @@ get_data_attr <- function(x, default = "data-id") {
 interactive_attr_toxml <- function(x,
                                    ids = character(0),
                                    rows = NULL) {
-  if (length(ids) < 1)
+  if (length(ids) < 1) {
     return(invisible())
+  }
 
   ip <- get_interactive_data(x)
   ipar <- get_ipar(x)
@@ -366,8 +367,9 @@ interactive_attr_toxml <- function(x,
   anames <- Filter(x = get_interactive_attr_names(ip, ipar = ipar), function(a) {
     !is.null(ip[[a]])
   })
-  if (length(anames) == 0)
+  if (length(anames) == 0) {
     return(invisible())
+  }
 
   for (a in anames) {
     if (length(rows) == 0) {
@@ -376,19 +378,24 @@ interactive_attr_toxml <- function(x,
       attrValue <- ip[[a]][rows]
     }
     attrValue <- switch(a,
-                        tooltip = encode_cr(attrValue),
-                        hover_css = check_css_attr(attrValue),
-                        selected_css = check_css_attr(attrValue),
-                        attrValue)
-    if (!is.character(attrValue))
+      tooltip = encode_cr(attrValue),
+      hover_css = check_css_attr(attrValue),
+      selected_css = check_css_attr(attrValue),
+      attrValue
+    )
+    if (!is.character(attrValue)) {
       attrValue <- as.character(attrValue)
+    }
     attrName <- switch(a,
-                       tooltip = "title",
-                       data_id = data_attr,
-                       a)
-    set_attr(name = attrName,
-             ids = as.integer(ids),
-             values = attrValue)
+      tooltip = "title",
+      data_id = data_attr,
+      a
+    )
+    set_attr(
+      name = attrName,
+      ids = as.integer(ids),
+      values = attrValue
+    )
   }
   invisible()
 }