+++ R/scale_interactive.R
@@ -1,5 +1,6 @@
 #' Calls a base scale function and returns an interactive scale.
 #' @noRd
+#' @importFrom rlang inherits_any
 scale_interactive <- function(scale_func,
                               ...,
                               extra_interactive_params = NULL) {
@@ -14,19 +15,20 @@ scale_interactive <- function(scale_func,
   if (isFALSE(sc$guide) || is.null(sc$guide)) {
     # no guide
     return(sc)
-
   } else if (is.character(sc$guide)) {
     if (sc$guide %in% c("none", "guide_none")) {
       # exit
       return(sc)
     } else if (sc$guide %in% c("legend", "bins", "colourbar", "colorbar", "coloursteps", "colorsteps")) {
       sc$guide <- paste0(sc$guide, "_interactive")
-    } else if (sc$guide %in% c("legend_interactive",
-                               "bins_interactive",
-                               "colourbar_interactive",
-                               "colorbar_interactive",
-                               "coloursteps_interactive",
-                               "colorsteps_interactive")) {
+    } else if (sc$guide %in% c(
+      "legend_interactive",
+      "bins_interactive",
+      "colourbar_interactive",
+      "colorbar_interactive",
+      "coloursteps_interactive",
+      "colorsteps_interactive"
+    )) {
       # ok
     } else {
       # the name could be a guide set by guides() and it might be interactive, but also it might not.
@@ -35,16 +37,23 @@ scale_interactive <- function(scale_func,
   } else if (inherits(sc$guide, "guide_none")) {
     # exit
     return(sc)
-  } else if (inherits(sc$guide, "interactive_guide")) {
+  } else if (inherits_any(sc$guide, c(
+    "GuideInteractiveLegend", "GuideInteractiveBins", "GuideInteractiveColourbar", "GuideInteractiveColoursteps"
+  ))
+  ) {
     # ok
-  } else if (inherits(sc$guide, "legend")) {
-    class(sc$guide) <- c("interactive_legend", "interactive_guide", class(sc$guide))
-  } else if (inherits(sc$guide, "bins")) {
-    class(sc$guide) <- c("interactive_bins", "interactive_guide", class(sc$guide))
-  } else if (inherits(sc$guide, "coloursteps") || inherits(sc$guide, "colorsteps")) {
-    class(sc$guide) <- c("interactive_coloursteps", "interactive_guide", class(sc$guide))
-  } else if (inherits(sc$guide, "colourbar") || inherits(sc$guide, "colorbar")) {
-    class(sc$guide) <- c("interactive_colourbar", "interactive_guide", class(sc$guide))
+  } else if (inherits_any(sc$guide, c(
+    "GuideLegend", "GuideBins", "GuideColourbar", "GuideColoursteps"
+  ))
+  ) {
+    interactive_guide <- NULL
+    if (inherits(sc$guide, "GuideColourbar")) {
+      # we must resolve if it's GuideColourbar or GuideColoursteps
+      if (!is.null(sc$guide$params$even.steps)) {
+        interactive_guide <- GuideInteractiveColoursteps
+      }
+    }
+    sc$guide <- guide_interactive(sc$guide, interactive_guide = interactive_guide)
   } else {
     warning("Only `legend`, 'bins', `colourbar` and `coloursteps` guides are supported for interactivity")
     return(sc)